requirements.txt
# LangChain - Exact versions for Phase 1-B freeze
langchain==0.3.26
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-community==0.3.27
langchain-chroma==0.2.3
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-checkpoint-sqlite==2.0.10

# Core Dependencies
numexpr==2.10.2
openai==1.93.0
chromadb==0.5.23
sentence-transformers==2.2.2
python-dotenv==1.0.1
pydantic==2.11.7
pydantic-core==2.33.2

# Security Notes:
# - chromadb 0.5.23 requires sentence-transformers <3; 2.2.2 satisfies that
# - transformers vulnerabilities remain; we suppress them via .pip-audit-ignore until upstream updates
# - torch 2.7.1 vulnerability also suppressed (GHSA-887c-mr87-cxwp)

# HTTP Client Dependencies
requests>=2.32.0           # HTTP client with certificate verification fixes
types-requests==2.32.4.20250611  # Type stubs for requests library
validators==0.28.1  # URL validation for security filtering

# Search API Dependencies
google-search-results==2.4.2  # SerpAPI client for web search functionality

# Web Content Extraction Dependencies (Phase 2 - Epic E-01)
trafilatura==1.12.2        # Primary extraction engine with enterprise reliability
selectolax==0.3.25         # High-performance HTML parsing for speed-critical scenarios  
lxml>=4.6.5                # Security-patched XML/HTML processing (CVE fixes)
beautifulsoup4>=4.12.0     # HTML/XML parsing library
types-beautifulsoup4==4.12.0.20250516  # Type stubs for BeautifulSoup
textstat>=0.7.8            # Text quality scoring and readability metrics

# Development Tools
gitpython>=3.1.41  # Security fix for PYSEC-2024-4, PYSEC-2022-42992, PYSEC-2023-137, PYSEC-2023-161, PYSEC-2023-165
yamllint
bandit>=1.8.6  # Security scanning tool
pytest-cov>=6.2.1  # Coverage testing for CI/CD
types-beautifulsoup4>=4.12.0.20250516  # Type stubs for BeautifulSoup

# Rich Console Support (for citation superscripts)
rich>=13.0.0  # Terminal formatting and Unicode superscript support

# Dashboard Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
websockets>=12.0
sse-starlette>=1.8.2,<3.0.0 